refactor(sanity): move string input into directory
This paves the way for adding a new string input implementation based
on Portable Text Editor.
