+++ packages/sanity/src/core/form/inputs/StringInput/StringInput.tsx
@@ -1,6 +1,6 @@
 import {TextInput} from '@sanity/ui'
 
-import {type StringInputProps} from '../types'
+import {type StringInputProps} from '../../types'
 
 /**
  *
