+++ packages/sanity/src/core/form/inputs/index.ts
@@ -18,7 +18,7 @@ export * from './PortableText/PortableTextInput'
 export {PortableTextInput as BlockEditor} from './PortableText/PortableTextInput'
 export * from './SelectInput'
 export * from './Slug/SlugInput'
-export * from './StringInput'
+export * from './StringInput/StringInput'
 export * from './TagsArrayInput'
 export * from './TelephoneInput'
 export * from './TextInput'
