+++ packages/sanity/src/core/form/studio/inputResolver/defaultInputs.ts
@@ -8,7 +8,7 @@ import {EmailInput} from '../../inputs/EmailInput'
 import {NumberInput} from '../../inputs/NumberInput'
 import {ObjectInput} from '../../inputs/ObjectInput'
 import {SlugInput} from '../../inputs/Slug/SlugInput'
-import {StringInput} from '../../inputs/StringInput'
+import {StringInput} from '../../inputs/StringInput/StringInput'
 import {TextInput} from '../../inputs/TextInput'
 import {UrlInput} from '../../inputs/UrlInput'
 import {StudioCrossDatasetReferenceInput} from '../inputs/crossDatasetReference/StudioCrossDatasetReferenceInput'
