+++ packages/sanity/src/core/form/studio/inputResolver/resolveStringInput.ts
@@ -1,7 +1,7 @@
 import {type StringSchemaType} from '@sanity/types'
 
 import {SelectInput} from '../../inputs/SelectInput'
-import {StringInput} from '../../inputs/StringInput'
+import {StringInput} from '../../inputs/StringInput/StringInput'
 import {getOption} from './helpers'
 
 export function resolveStringInput(type: StringSchemaType) {
